finish moving Country entity to geoprocessingDB connection
