+++ api/src/modules/countries/countries.module.ts
@@ -2,11 +2,11 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
 import { CountriesController } from './countries.controller';
-import { Country } from './country.api.entity';
+import { Country } from './country.geo.entity';
 import { CountriesService } from './countries.service';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Country])],
+  imports: [TypeOrmModule.forFeature([Country], 'geoprocessingDB')],
   providers: [CountriesService],
   controllers: [CountriesController],
   exports: [CountriesService],
