+++ api/test/countries.e2e-spec.ts
@@ -3,7 +3,7 @@ import { INestApplication } from '@nestjs/common';
 import * as request from 'supertest';
 import { AppModule } from './../src/app.module';
 import { E2E_CONFIG } from './e2e.config';
-import { Country } from 'modules/countries/country.api.entity';
+import { Country } from 'modules/countries/country.geo.entity';
 
 describe('CountriesModule (e2e)', () => {
   let app: INestApplication;
